Revert "Mvvm branch"
