+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,16 @@
         android:theme="@style/Theme.RadioApp"
         tools:targetApi="31">
         <activity
-            android:name=".view.MainActivity"
+            android:name=".MainActivity2"
+            android:exported="true">
+
+
+            <meta-data
+                android:name="android.app.lib_name"
+                android:value="" />
+        </activity>
+        <activity
+            android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
